refactor: flatten register request structure
